Fixes in Tests in supplier_management_site, New Organization in Test in invoices_app
